Add files via upload
This code plots an animation of two double pendulums swinging on the same axes. This serves to provide a visual comparison for what happens when two double pendulums are released at slightly different angles. In other words it shows a visual representation of how sensitive a double pendulum is to initial conditions.